sprint 3 commit

+++ src/main/java/com/ust/pos/dto/PriceDto.java
@@ -3,11 +3,13 @@ package com.ust.pos.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class PriceDto extends CommonDto {
 
-    private Double costPrice;
+    private BigDecimal costPrice;
     private String product;
     private String priceType;
 }
