+++ src/main/java/com/ust/pos/model/Price.java
@@ -4,12 +4,14 @@ import jakarta.persistence.Entity;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Entity
 public class Price extends CommonFields {
 
-    private Double costPrice;
+    private BigDecimal costPrice;
     private String product;
     private String priceType;
 }
